STAC-24722: optionally clear stackpacks

+++ docker/rootfs/copy-stackpacks.sh
@@ -5,12 +5,14 @@ set -e
 target_dir=$1
 
 if [ -z "${target_dir}" ]; then
-  echo "Usage: $0 <target_dir>"
+  echo "Usage: $0 <target_dir> [--clear]"
   exit 1
 fi
 
-echo "Cleaning up current Stackpacks in ${target_dir}/"
-rm -rf "${target_dir:?}/"* || true
+if [[ $# -ge 2 && $2 == "--clear" ]]; then
+  echo "Cleaning up current Stackpacks in ${target_dir}/"
+  rm -rf "${target_dir:?}/"* || true
+fi
 
 echo "Copying Stackpacks from /stackpacks to ${target_dir}/"
